+++ Assets/Prefabs/SlimeRabbit_Prefab.prefab
@@ -146,6 +146,7 @@ GameObject:
   - component: {fileID: 2643115779267335905}
   - component: {fileID: 2643115779267335906}
   - component: {fileID: 2643115779267335908}
+  - component: {fileID: 4202900052421039268}
   m_Layer: 0
   m_Name: SlimeRabbit_Prefab
   m_TagString: Untagged
@@ -257,6 +258,28 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 61b38d7e12126f84e87c156a80bf8ddb, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+--- !u!114 &4202900052421039268
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2641491944515610799}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 0566e554e0956bf44a0c9a87933fac0a, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  aiState: 0
+  enemyData:
+    id: 0
+    name: SlimeRabit
+    maxHp: 3
+    attackPower: 1
+    moveSpeed: 5
+    attackInterval: 5
+    attackRangeType: 0
+  shootDistance: 0
 --- !u!1 &2641512759266206003
 GameObject:
   m_ObjectHideFlags: 0
@@ -748,8 +771,8 @@ GameObject:
   m_Component:
   - component: {fileID: 2643115779438074872}
   - component: {fileID: 2643115779438074870}
-  - component: {fileID: 2643115779438074873}
   - component: {fileID: 2643115779438074875}
+  - component: {fileID: 6601553247401444777}
   m_Layer: 0
   m_Name: Anchor
   m_TagString: Untagged
@@ -788,29 +811,6 @@ Rigidbody:
   m_Interpolate: 0
   m_Constraints: 112
   m_CollisionDetection: 0
---- !u!114 &2643115779438074873
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 2643115779438074871}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0566e554e0956bf44a0c9a87933fac0a, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  moveTime: 5
-  shootDistance: 0
-  aiState: 0
-  enemyData:
-    id: 0
-    name: 
-    maxHp: 0
-    attackPower: 0
-    moveSpeed: 0
-    attackInterval: 0
-    attackRangeType: 0
 --- !u!135 &2643115779438074875
 SphereCollider:
   m_ObjectHideFlags: 0
@@ -824,3 +824,16 @@ SphereCollider:
   serializedVersion: 2
   m_Radius: 0.5
   m_Center: {x: 0, y: 0.45, z: 0}
+--- !u!114 &6601553247401444777
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2643115779438074871}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 29794af0c8937be4ea822f040aa8a115, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  moveTime: 5
