+++ Assets/Scripts/EnemyMove.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 29794af0c8937be4ea822f040aa8a115
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
